精简代码结构
