+++ README.md
@@ -24,16 +24,16 @@
    ```yaml
    # 邮件提醒服务设置，如果不需要， 将email下的字段设置为null即可
    email:
-     from: "12345@qq.com"                    # 发送邮件的地址
-     username: "12345@qq.com"	       # 登陆邮箱的账号
-     password: "QQmima"	                   # smtp服务授权码，请进入邮箱设置进行查看
-     smtp: "smtp.qq.com"						 # smtp服务器
-     port: 465                                                  # smtp服务器SSL协议端口号，大多数邮箱默认为465
+     from: "12345@qq.com" # 发送邮件的地址
+     username: "12345@qq.com" # 登陆邮箱的账号
+     password: "QQmima" # smtp服务授权码，请进入邮箱设置进行查看
+     smtp: "smtp.qq.com" # smtp服务器
+     port: 465 # smtp服务器SSL协议端口号，大多数邮箱默认为465
    
    users:
-     - id: "user1"	                                           # 学号
-       pwd: "pwd"	                                        # 密码
-       email_to: "user1@qq.com"	      # 接受报送成功邮件的邮箱地址
+     - id: "user1" # 学号
+       pwd: "pwd"	# 密码
+       email_to: "user1@qq.com"	# 接受报送成功邮件的邮箱地址
      - id: "user2"
        pwd: "pwd"
        email_to: "user2@qq.com"
@@ -76,14 +76,14 @@
    - 测试邮箱服务是否可使用
 
      ```python
-     python main.py -e -a xxx@xx.com					# xxx@xx.com为接收测试邮件的账号
+     python main.py -e -a xxx@xx.com # xxx@xx.com为接收测试邮件的账号
      ```
 
 4. 启动程序：
 
    ```python
-   nohup python main.py > shu_report.log 2>&1 & 		# 针对Linux，启动程序，后台运行，输出结果导出shu_report.log
-   																											  # 如需设置开机自启动，请自行搜索网上教程
+   nohup python main.py > shu_report.log 2>&1 & # 针对Linux，启动程序，后台运行，输出结果导出shu_report.log
+   # 如需设置开机自启动，请自行搜索网上教程
    ```
 
 ## 要求
